+++ composeApp/src/commonMain/kotlin/com/blankthings/baseapp/data/NoteRepository.kt
@@ -7,5 +7,5 @@ interface NoteRepository {
     suspend fun addNote(note: Note)
     suspend fun updateNote(note: Note)
     suspend fun deleteNote(note: Note)
-    fun getNote(noteId: Int): Note
+    fun getNote(noteId: Long): Note
 }
